Language.Haskell.TH.Instances

Description

It provides the following instances:

Ord, Lift, Generic, Show, Eq, Data, Typeable, Ppr, instances for the datatypes in Language.Haskell.TH

Lift instances for Data.Word / Data.Int types

Applicative for Q

Quasi for ReaderT, WriterT, StateT, and RWST.

Typeable for Lift, NameIs, Ppr, PprM, Q, Quasi, QuasiQuoter, and TExp

More recent versions of template-haskell, particularly 2.10 (GHC 7.10), provide these instances. However, in order to support older versions you should import this module.

Note that the Ord instances are not guaranteed to produce consistent results across template-haskell / GHC versions, as they have different data types, with different constructor orders.

This module also implicitly re-exports the instances defined in Instances.TH.Lift. This is mostly to ensure that there aren't collisions of orphans between th-orphans and th-lift-instances.

Orphan instances

Lift Exp

Methods

lift :: Exp -> Q Exp

Lift Match

Methods

lift :: Match -> Q Exp

Lift Clause

Methods

lift :: Clause -> Q Exp

Lift Pat

Methods

lift :: Pat -> Q Exp

Lift Type

Methods

lift :: Type -> Q Exp

Lift Dec

Methods

lift :: Dec -> Q Exp

Lift FunDep

Methods

lift :: FunDep -> Q Exp

Lift TyVarBndr

Methods

lift :: TyVarBndr -> Q Exp

Lift InjectivityAnn

Methods

lift :: InjectivityAnn -> Q Exp

Lift Overlap

Methods

lift :: Overlap -> Q Exp

Lift DerivStrategy

Methods

lift :: DerivStrategy -> Q Exp

Lift Loc

Methods

lift :: Loc -> Q Exp

Lift Info

Methods

lift :: Info -> Q Exp

Lift Fixity

Methods

lift :: Fixity -> Q Exp

Lift FixityDirection

Methods

lift :: FixityDirection -> Q Exp

Lift Lit

Methods

lift :: Lit -> Q Exp

Lift Body

Methods

lift :: Body -> Q Exp

Lift Guard

Methods

lift :: Guard -> Q Exp

Lift Stmt

Methods

lift :: Stmt -> Q Exp

Lift Range

Methods

lift :: Range -> Q Exp

Lift DerivClause

Methods

lift :: DerivClause -> Q Exp

Lift TypeFamilyHead

Methods

lift :: TypeFamilyHead -> Q Exp

Lift TySynEqn

Methods

lift :: TySynEqn -> Q Exp

Lift Foreign

Methods

lift :: Foreign -> Q Exp

Lift Callconv

Methods

lift :: Callconv -> Q Exp

Lift Safety

Methods

lift :: Safety -> Q Exp

Lift Pragma

Methods

lift :: Pragma -> Q Exp

Lift Inline

Methods

lift :: Inline -> Q Exp

Lift RuleMatch

Methods

lift :: RuleMatch -> Q Exp

Lift Phases

Methods

lift :: Phases -> Q Exp

Lift RuleBndr

Methods

lift :: RuleBndr -> Q Exp

Lift AnnTarget

Methods

lift :: AnnTarget -> Q Exp

Lift SourceUnpackedness

Methods

lift :: SourceUnpackedness -> Q Exp

Lift SourceStrictness

Methods

lift :: SourceStrictness -> Q Exp

Lift Con

Methods

lift :: Con -> Q Exp

Lift Bang

Methods

lift :: Bang -> Q Exp

Lift PatSynDir

Methods

lift :: PatSynDir -> Q Exp

Lift PatSynArgs

Methods

lift :: PatSynArgs -> Q Exp

Lift FamilyResultSig

Methods

lift :: FamilyResultSig -> Q Exp

Lift TyLit

Methods

lift :: TyLit -> Q Exp

Lift Role

Methods

lift :: Role -> Q Exp

Quasi m => Quasi (StateT s m)

Methods

qNewName :: String -> StateT s m Name

qReport :: Bool -> String -> StateT s m ()

qRecover :: StateT s m a -> StateT s m a -> StateT s m a

qLookupName :: Bool -> String -> StateT s m (Maybe Name)

qReify :: Name -> StateT s m Info

qReifyFixity :: Name -> StateT s m (Maybe Fixity)

qReifyInstances :: Name -> [Type] -> StateT s m [Dec]

qReifyRoles :: Name -> StateT s m [Role]

qReifyAnnotations :: Data a => AnnLookup -> StateT s m [a]

qReifyModule :: Module -> StateT s m ModuleInfo

qReifyConStrictness :: Name -> StateT s m [DecidedStrictness]

qLocation :: StateT s m Loc

qRunIO :: IO a -> StateT s m a

qAddDependentFile :: FilePath -> StateT s m ()

qAddTopDecls :: [Dec] -> StateT s m ()

qAddForeignFile :: ForeignSrcLang -> String -> StateT s m ()

qAddModFinalizer :: Q () -> StateT s m ()

qGetQ :: Typeable * a => StateT s m (Maybe a)

qPutQ :: Typeable * a => a -> StateT s m ()

qIsExtEnabled :: Extension -> StateT s m Bool

qExtsEnabled :: StateT s m [Extension]

(Quasi m, Monoid w) => Quasi (WriterT w m)

Methods

qNewName :: String -> WriterT w m Name

qReport :: Bool -> String -> WriterT w m ()

qRecover :: WriterT w m a -> WriterT w m a -> WriterT w m a

qLookupName :: Bool -> String -> WriterT w m (Maybe Name)

qReify :: Name -> WriterT w m Info

qReifyFixity :: Name -> WriterT w m (Maybe Fixity)

qReifyInstances :: Name -> [Type] -> WriterT w m [Dec]

qReifyRoles :: Name -> WriterT w m [Role]

qReifyAnnotations :: Data a => AnnLookup -> WriterT w m [a]

qReifyModule :: Module -> WriterT w m ModuleInfo

qReifyConStrictness :: Name -> WriterT w m [DecidedStrictness]

qLocation :: WriterT w m Loc

qRunIO :: IO a -> WriterT w m a

qAddDependentFile :: FilePath -> WriterT w m ()

qAddTopDecls :: [Dec] -> WriterT w m ()

qAddForeignFile :: ForeignSrcLang -> String -> WriterT w m ()

qAddModFinalizer :: Q () -> WriterT w m ()

qGetQ :: Typeable * a => WriterT w m (Maybe a)

qPutQ :: Typeable * a => a -> WriterT w m ()

qIsExtEnabled :: Extension -> WriterT w m Bool

qExtsEnabled :: WriterT w m [Extension]

Quasi m => Quasi (ReaderT * r m)

Methods

qNewName :: String -> ReaderT * r m Name

qReport :: Bool -> String -> ReaderT * r m ()

qRecover :: ReaderT * r m a -> ReaderT * r m a -> ReaderT * r m a

qLookupName :: Bool -> String -> ReaderT * r m (Maybe Name)

qReify :: Name -> ReaderT * r m Info

qReifyFixity :: Name -> ReaderT * r m (Maybe Fixity)

qReifyInstances :: Name -> [Type] -> ReaderT * r m [Dec]

qReifyRoles :: Name -> ReaderT * r m [Role]

qReifyAnnotations :: Data a => AnnLookup -> ReaderT * r m [a]

qReifyModule :: Module -> ReaderT * r m ModuleInfo

qReifyConStrictness :: Name -> ReaderT * r m [DecidedStrictness]

qLocation :: ReaderT * r m Loc

qRunIO :: IO a -> ReaderT * r m a

qAddDependentFile :: FilePath -> ReaderT * r m ()

qAddTopDecls :: [Dec] -> ReaderT * r m ()

qAddForeignFile :: ForeignSrcLang -> String -> ReaderT * r m ()

qAddModFinalizer :: Q () -> ReaderT * r m ()

qGetQ :: Typeable * a => ReaderT * r m (Maybe a)

qPutQ :: Typeable * a => a -> ReaderT * r m ()

qIsExtEnabled :: Extension -> ReaderT * r m Bool

qExtsEnabled :: ReaderT * r m [Extension]

(Quasi m, Monoid w) => Quasi (RWST r w s m)

Methods

qNewName :: String -> RWST r w s m Name

qReport :: Bool -> String -> RWST r w s m ()

qRecover :: RWST r w s m a -> RWST r w s m a -> RWST r w s m a

qLookupName :: Bool -> String -> RWST r w s m (Maybe Name)

qReify :: Name -> RWST r w s m Info

qReifyFixity :: Name -> RWST r w s m (Maybe Fixity)

qReifyInstances :: Name -> [Type] -> RWST r w s m [Dec]

qReifyRoles :: Name -> RWST r w s m [Role]

qReifyAnnotations :: Data a => AnnLookup -> RWST r w s m [a]

qReifyModule :: Module -> RWST r w s m ModuleInfo

qReifyConStrictness :: Name -> RWST r w s m [DecidedStrictness]

qLocation :: RWST r w s m Loc

qRunIO :: IO a -> RWST r w s m a

qAddDependentFile :: FilePath -> RWST r w s m ()

qAddTopDecls :: [Dec] -> RWST r w s m ()

qAddForeignFile :: ForeignSrcLang -> String -> RWST r w s m ()

qAddModFinalizer :: Q () -> RWST r w s m ()

qGetQ :: Typeable * a => RWST r w s m (Maybe a)

qPutQ :: Typeable * a => a -> RWST r w s m ()

qIsExtEnabled :: Extension -> RWST r w s m Bool

qExtsEnabled :: RWST r w s m [Extension]